Decorate SolrDocument + add test for browse_controller
