+++ spec/controllers/browse_controller_spec.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+require 'rails_helper'
+
+RSpec.describe BrowseController, type: :controller do
+  describe 'GET #index' do
+    subject { response }
+
+    before { get :index }
+
+    it { is_expected.to be_successful }
+
+    it 'builds a shelf list' do
+      expect(assigns(:shelf_list)).to be_a(ShelfListPresenter)
+    end
+  end
+end
